+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this module are documented in this file.
 
 poloniex-api-node module adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.4.0] - 2017-09-02
+- Added: `limit` parameter for `returnTradeHistory` and `returnMyTradeHistory` methods
+- Added: contributors to [README.md](README.md#contributors) (give credit where credit is due)  
+
 ## [1.3.1] - 2017-07-11
 - Fixed: throw error when invalid parameters (#2)